# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Second-hand Trading Platform API
  version: 0.1.0
servers:
  - url: http://localhost:7001
paths:
...
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: 获取当前登录用户信息
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 当前用户信息
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: 未登录或登录已过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/items/{id}/reviews:
    get:
      operationId: listItemReviews
      summary: 获取商品评价列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 评价列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
        "400":
          description: 无效的商品 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/reviews:
    get:
      operationId: listPendingReviews
      summary: 查看待审核商品列表（仅管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 待审核商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 仅管理员可执行此操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/reviews:
    post:
      operationId: createReview
      summary: 提交评价（需认证，仅已签收买家）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: 评价创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Review"
        "400":
          description: 参数错误或评分无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权评价此订单
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 已评价过
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sellers/{id}/reviews:
    get:
      operationId: listSellerReviews
      summary: 获取商家近期评价（预留）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商家 ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            default: 30
          description: 查询近多少天的评价
      responses:
        "200":
          description: 评价列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
        "400":
          description: 无效的商家 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商家不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
  schemas:
    User:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Item:
      type: object
      required:
        [
          id,
          title,
          price,
          quantity,
          description,
          images,
          coverImage,
          category,
          sellerId,
          sellerName,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        title:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        coverImage:
          type: string
        category:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        status:
          type: string
          enum: [pending, active, delisted]
        createdAt:
          type: string
          format: date-time
        quantityUpdatedAt:
          type: string
          format: date-time
          nullable: true
        isItemFavorited:
          type: boolean
          description: 当前用户是否收藏该商品（仅登录用户返回有效值）
        isSellerFavorited:
          type: boolean
          description: 当前用户是否收藏该商家（仅登录用户返回有效值）
    ItemListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        totalPages:
          type: integer
...
    Review:
      type: object
      required:
        [
          id,
          userId,
          orderId,
          itemId,
          sellerId,
          rating,
          comment,
          username,
          createdAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        orderId:
          type: integer
        itemId:
          type: integer
        sellerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReviewInput:
      type: object
      additionalProperties: false
      required: [orderId, itemId, rating]
      properties:
        orderId:
          type: integer
        itemId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
          description: 评论文本（可选）
    ReviewListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        totalPages:
          type: integer